refactorの間違いを直した
